+++ src/Modules/Tours/Explorer.Tours.API/Public/Administration/ITourExecutionService.cs
@@ -21,7 +21,7 @@ namespace Explorer.Tours.API.Public.Administration
         Result StartNewTour(TourExecutionDto tourExecution);
 
         Result<TourExecutionDto> LeaveTour(int touristId, int tourId);
-        Result<TourExecutionDto> CheckLocation(int id, double latitude, double longitude);
+        Result<TourExecutionDto> CheckLocation(int id, float latitude, float longitude);
         Result<List<KeyPointDto>> GetCompletedKeyPoints(int id);
 
     }
